[IMP] Refactoring models
