+++ formio_component/models/formio_component.py
@@ -61,104 +61,3 @@ class FormioComponent(models.Model):
                 )
             res.append((r.id, label))
         return res
-
-    # ----------------------------------------------------------
-    # Helper
-    # ----------------------------------------------------------
-
-    def _get_component(self, builder, comp_id):
-        """
-        Returns a formio.component obj from component_id.
-
-        :param tuple builder: the builder records to which the component belongs,
-        :param string comp_id: Component id of desired component in database,
-        :return Database record according to builder_ids and comp_id.
-        """
-        return self.search([
-            ("builder_id", 'in', builder.ids),
-            ("component_id", '=', comp_id)
-        ])
-
-    # ----------------------------------------------------------
-    # Public
-    # ----------------------------------------------------------
-
-    def compare_components(self, builder):
-        """
-        Compares arrays with component keys.
-
-        :param tuple builder: the builder records to which the components belong,
-        :return dict: with the deleted and added component_ids.
-
-        {   'added': ['el1t2b', 'e2zvxh4', ...],
-            'deleted': ['ejswb', 'ema9axd', ...]
-        }
-
-        """
-        new_components = []
-        old_components = []
-        for comp_id, obj in builder._formio.component_ids.items():
-            new_components.append(comp_id)
-        for record in self.search([("builder_id", 'in', builder.ids)]):
-            old_components.append(record.component_id)
-        return {
-            'added': list(set(new_components).difference(old_components)),
-            'deleted': list(set(old_components).difference(new_components))
-        }
-
-    def write_components(self, builder, comp_ids):
-        """
-        Writes the components with all required data to formio.component model.
-
-        :param tuple builder: the builder where the components are located,
-        :param array component_keys: components which should be added to this model.
-        """
-        for comp_id in comp_ids:
-            obj = builder._formio.component_ids[comp_id]
-            self.create({
-                'label': obj.label,
-                'component_id': obj.id,
-                'key': obj.key,
-                'type': obj.type,
-                'builder_id': builder.id,
-            })
-
-    def update_components(self, builder):
-        """
-        Checks for any component related changes and synchronize them with database records.
-
-        :param builder: The builder where the components are located at.
-        """
-        for comp_id, obj in builder._formio.component_ids.items():
-            record = self._get_component(builder, comp_id)
-            if not record or record.component_id != obj.id:
-                continue
-
-            """
-            Updating component attributes
-            """
-            if record.label != obj.label:
-                record.label = obj.label
-            if record.key != obj.key:
-                record.key = obj.key
-
-            """
-            Updating datagrid
-            """
-            if obj.parent:
-                if record.parent_id.component_id != obj.parent.id:
-                    parent_record = self._get_component(builder, obj.parent.id)
-                    record.parent_id = parent_record
-            elif not obj.parent and record.parent_id:
-                record.parent_id = False
-
-    def delete_components(self, builder, comp_ids):
-        """
-        Removes components from formio.component model.
-
-        :param tuple builder: the builder record where the components are located,
-        :param array component_keys: components which should be removed from this model.
-        """
-        for comp_id in comp_ids:
-            components = self._get_component(builder, comp_id)
-            components.unlink()
