+++ formio_data_api/models/formio_builder.py
@@ -6,7 +6,7 @@ import logging
 from formiodata.builder import Builder
 
 from odoo import fields, models, _
-from odoo.exceptions import ValidationError
+from odoo.exceptions import ValidationError, UserError
 
 
 _logger = logging.getLogger(__name__)
@@ -18,6 +18,39 @@ class FormioBuilder(models.Model):
     component_server_api_ids = fields.One2many(
         'formio.component.server.api', 'formio_builder_id',
         string='Component Server APIs', context={'active_test': False})
+
+    def __getattr__(self, name):
+        if name == '_formio' and self._name == 'formio.builder':
+            # TODO implement caching on the model object
+            # self._cache or self.env.cache API only works for model fields, not Python attr.
+
+            # if '_formio' not in self.__dict__:
+            no_cache = True
+            if no_cache:
+                context = self._context
+                if 'lang' in context:
+                    lang = context['lang']
+                elif 'lang' not in context and 'uid' in context:
+                    lang = self.env['res.users'].browse(context['uid']).lang
+                elif 'lang' not in context and 'uid' not in context:
+                    lang = self.write_uid.lang
+                else:
+                    raise UserError("The form can't be loaded. No (user) language was set.")
+
+                res_lang = self.env['res.lang'].search([('code', '=', lang)], limit=1)
+
+            if self.schema is False:
+                # HACK masquerade empty Builder object
+                builder_obj = Builder('{}')
+            else:
+                builder_obj = Builder(
+                    self.schema,
+                    language=res_lang.iso_code,
+                    i18n=self.i18n_translations()
+                )
+            return builder_obj
+        else:
+            return self.__getattribute__(name)
 
     def _component_api_keys(self):
         return ['model_field', 'res_field', 'user_field', 'code_api']
